\name{touchard-package}
\alias{touchard-package}
\docType{package}
\title{
\packageTitle{touchard}
}
\description{
%\packageDescription{touchard}
Analysis of count data with the Touchard model including: d-*, p-*, q-* and r-* functions; 
inference for univariate analysis 
(maximum likelihood and method of moment estimation); regression modeling
and graphical assessment of goodness of fit (rootogram and touchardness plot).
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{touchard}
\packageIndices{touchard}

}
\author{
\packageAuthor{touchard}

Maintainer: \packageMaintainer{touchard}
}
\references{

Matsushita RY, Pianto D, Andrade BB, Cancado A, Silva S (2018) The Touchard distribution, 
\emph{Communications in Statistics - Theory and Methods}, <doi:10.1080/03610926.2018.1444177>

	
Andrade, BB; Matsushita, RY; Oliveira, SB (submitted) Analyzing Count Data with the Touchard Model in R. 
\emph{J. of Statistical Software}.
}



